conversational-document-rag-using-ollama-groq/requirement.txt
langchain
langchain_community
streamlit
python-dotenv
langchain_ollama
langchain_groq
langchain-text-splitters
chromadb
pypdf